Replace Firebase sign-in with Google sign-in
