+++ dogfooding/src/main/kotlin/io/getstream/video/android/ui/login/LoginScreen.kt
@@ -62,9 +62,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
+import androidx.core.content.ContextCompat.getString
 import androidx.core.content.ContextCompat.startActivity
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.firebase.ui.auth.AuthUI
+import com.google.android.gms.auth.api.signin.GoogleSignIn
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions
+import com.google.android.gms.common.api.Scope
 import io.getstream.video.android.BuildConfig
 import io.getstream.video.android.R
 import io.getstream.video.android.compose.theme.VideoTheme
@@ -72,6 +75,7 @@ import io.getstream.video.android.ui.theme.Colors
 import io.getstream.video.android.ui.theme.LinkText
 import io.getstream.video.android.ui.theme.LinkTextData
 import io.getstream.video.android.ui.theme.StreamButton
+
 
 @Composable
 fun LoginScreen(
@@ -283,13 +287,15 @@ private fun HandleLoginUiStates(
     LaunchedEffect(key1 = loginUiState) {
         when (loginUiState) {
             is LoginUiState.GoogleSignIn -> {
-                val providers = arrayListOf(
-                    AuthUI.IdpConfig.GoogleBuilder().build(),
-                )
-
-                val signInIntent = AuthUI.getInstance().createSignInIntentBuilder()
-                    .setAvailableProviders(providers)
+                val gso = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+                    .requestEmail()
+                    .requestIdToken(getString(context, R.string.default_web_client_id))
+                    .requestServerAuthCode(getString(context, R.string.default_web_client_id))
+                    .requestScopes(Scope("https://www.googleapis.com/auth/directory.readonly"))
                     .build()
+                val gsc = GoogleSignIn.getClient(context, gso)
+
+                val signInIntent: Intent = gsc.signInIntent
                 signInLauncher.launch(signInIntent)
             }
